requirements.txt
gradio
transformers
qwen-vl-utils
torch
Pillow
